Add a dedicated JS file containing datasets-related functions
