+++ frontend/src/src/components/DatasetSelection.svelte
@@ -1,14 +1,14 @@
 <script>
   import {onMount, createEventDispatcher, afterUpdate} from 'svelte';
-  import {getDatasetsWithVariables, getDatasetLayerName} from '../client.js';
-  import {createLayer} from '../layers.js';
+  import {datasetsStore, datasetTopicsStore} from '../stores.js';
+  import {getDatasetLayerName} from '../client.js';
+  import {getDatasets, getDatasetBySharedId, getDataset} from '../datasets.js';
+  import {createLayerSimple} from '../layers.js';
   import {BASE_URL} from '../settings.js';
 
 
-  let availableDatasets = null;
   let filter = '';
   let filteredDatasets = [];
-  const topics = [];
   let selectedTopic = null;
 
   let rootElement = null;
@@ -20,69 +20,7 @@
 
 
   onMount(async () => {
-    const datasets = await getDatasetsWithVariables();
-
-    for (const dataset of datasets) {
-      dataset.open = false;
-
-      dataset.info.both = false;
-      dataset.info.variables_only = false;
-      dataset.info.time_period_only = false;
-      dataset.info.unique = false;
-      dataset.info.const_variable = null;
-      dataset.info.const_time_period = null;
-      dataset.info.open_intermediate_layers = [];
-
-      if ((dataset.info.variables.length > 1) && (dataset.info.time_periods.length > 1)) {
-        dataset.info.both = true;
-      } else if ((dataset.info.variables.length > 1) &&
-                 (dataset.info.time_periods.length == 1)) {
-        dataset.info.variables_only = true;
-        dataset.info.const_time_period = dataset.info.time_periods[0];
-      } else if ((dataset.info.variables.length > 1) &&
-                 (dataset.info.time_periods.length == 0)) {
-        dataset.info.variables_only = true;
-      } else if ((dataset.info.variables.length == 1) &&
-                 (dataset.info.time_periods.length > 1)) {
-        dataset.info.time_period_only = true;
-        dataset.info.const_variable = dataset.info.variables[0];
-      } else if ((dataset.info.variables.length == 0) &&
-                 (dataset.info.time_periods.length > 1)) {
-        dataset.info.time_period_only = true;
-      } else {
-        dataset.info.unique = true;
-
-        if (dataset.info.variables.length == 1) {
-          dataset.info.const_variable = dataset.info.variables[0];
-        }
-
-        if (dataset.info.time_periods.length == 1) {
-          dataset.info.const_time_period = dataset.info.time_periods[0];
-        }
-      }
-
-      if ((dataset.group != '') && (topics.indexOf(dataset.group) == -1)) {
-        topics.push(dataset.group);
-      }
-    }
-
-    datasets.sort(function(dataset1, dataset2) {
-      const title1 = dataset1.title;
-      const title2 = dataset2.title;
-      if (title1 < title2) {
-        return -1;
-      }
-      if (title1 > title2) {
-        return 1;
-      }
-      return 0;
-    });
-
-    topics.sort();
-
-    console.log(datasets.length + ' datasets found, in ' + topics.length + ' topics');
-
-    availableDatasets = datasets;
+    await getDatasets();
 
     // Allow to display a dataset at launch
     let startDataset = null;
@@ -90,21 +28,9 @@
     const urlParams = new URLSearchParams(window.location.search);
 
     if (urlParams.has('shared_id')) {
-      const sharedId = urlParams.get('shared_id');
-      for (const dataset of datasets) {
-        if (dataset.shared_id == sharedId) {
-          startDataset = dataset;
-          break;
-        }
-      }
+      startDataset = getDatasetBySharedId(urlParams.get('shared_id'));
     } else if (urlParams.has('dataset_id')) {
-      const datasetId = parseInt(urlParams.get('dataset_id'));
-      for (const dataset of datasets) {
-        if (dataset.ds_id == datasetId) {
-          startDataset = dataset;
-          break;
-        }
-      }
+      startDataset = getDataset(urlParams.get('dataset_id'));
     }
 
     if (startDataset !== null) {
@@ -121,7 +47,7 @@
         startTimePeriod = startDataset.info.time_periods[0];
       }
 
-      addLayer(startDataset.ds_id, startVariable, startTimePeriod);
+      createLayerSimple(startDataset.ds_id, startVariable, startTimePeriod);
     }
   });
 
@@ -136,8 +62,8 @@
 
 
   $: {
-    if (availableDatasets !== null) {
-      filteredDatasets = Array.from(availableDatasets);
+    if ($datasetsStore !== null) {
+      filteredDatasets = Array.from($datasetsStore);
 
       if (selectedTopic !== null) {
         filteredDatasets = filteredDatasets.filter((dataset) =>
@@ -154,7 +80,7 @@
 
   async function toggleDataset(dataset) {
     dataset.open = !dataset.open;
-    availableDatasets = availableDatasets;
+    $datasetsStore = $datasetsStore;
   }
 
 
@@ -167,7 +93,7 @@
       dataset.info.open_intermediate_layers.push(variable);
     }
 
-    availableDatasets = availableDatasets;
+    $datasetsStore = $datasetsStore;
   }
 
 
@@ -176,54 +102,8 @@
   }
 
 
-  async function addLayer(datasetId, variable, timePeriod) {
-    const dataset = availableDatasets.filter((dataset) => dataset.ds_id == datasetId)[0];
-
-    const layerName = await getDatasetLayerName(
-        datasetId, dataset.is_raster, variable, timePeriod,
-    );
-
-    const labels = {
-      primary: null,
-      secondary: null,
-      dataset: null,
-    };
-
-    let title = null;
-
-    if (variable !== null) {
-      labels.primary = variable;
-      labels.dataset = dataset.title;
-
-      if (timePeriod !== null) {
-        labels.secondary = timePeriod;
-        title = variable + '\n\n' + timePeriod + '\n\n' + dataset.title;
-      } else {
-        title = variable + '\n\n' + dataset.title;
-      }
-    } else if (timePeriod !== null) {
-      labels.primary = timePeriod;
-      labels.dataset = dataset.title;
-      title = timePeriod + '\n\n' + dataset.title;
-    } else {
-      labels.primary = dataset.title;
-      title = dataset.title;
-    }
-
-    const layerInfos = {
-      dataset: dataset.ds_id,
-      variable: variable,
-    };
-
-    createLayer(
-        layerName, labels, title, dataset.is_raster, dataset.is_tiled,
-        dataset.info.min_zoom_level, null, layerInfos,
-    );
-  }
-
-
   async function copyGeoJSONUrlToClipboard(datasetId, variable, timePeriod, event) {
-    const dataset = availableDatasets.filter((dataset) => dataset.ds_id == datasetId)[0];
+    const dataset = $datasetsStore.filter((dataset) => dataset.ds_id == datasetId)[0];
 
     if (dataset.is_raster) {
       return;
@@ -420,17 +300,17 @@
      on:dblclick|stopPropagation on:wheel|stopPropagation>
   <h3>Datasets</h3>
 
-  {#if !availableDatasets}
+  {#if !$datasetsStore}
     Loading datasets...
   {:else}
     <input bind:value={filter} class="search" placeholder="Search dataset...">
 
-    {#if topics.length > 0}
+    {#if $datasetTopicsStore.length > 0}
       <div class="topics">
         <label for="topic">Group:</label>
         <select name="topic" bind:value={selectedTopic}>
           <option value={null}>All</option>
-          {#each topics as topic}
+          {#each $datasetTopicsStore as topic}
             <option value={topic}>{topic}</option>
           {/each}
         </select>
@@ -467,7 +347,7 @@
                       {#each dataset.info.time_periods as timePeriod}
                         {#if isCombinationValid(dataset, variable, timePeriod)}
                           <tr class="layer final" title={timePeriod}
-                              on:click={() => addLayer(dataset.ds_id, variable, timePeriod)}>
+                              on:click={() => createLayerSimple(dataset.ds_id, variable, timePeriod)}>
                             <td></td>
                             <td class="bullet"></td>
                             <td class="bullet">◦</td>
@@ -490,7 +370,7 @@
                 {:else if dataset.info.variables_only}
                   {#each dataset.info.variables as variable}
                     <tr class="layer final" title={variable}
-                        on:click={() => addLayer(dataset.ds_id, variable, dataset.info.const_time_period)}>
+                        on:click={() => createLayerSimple(dataset.ds_id, variable, dataset.info.const_time_period)}>
                       <td></td>
                       <td class="bullet">◦</td>
                       <td colspan="2">{variable}</td>
@@ -510,7 +390,7 @@
                 {:else if dataset.info.time_period_only}
                   {#each dataset.info.time_periods as timePeriod}
                     <tr class="layer final" title={timePeriod}
-                        on:click={() => addLayer(dataset.ds_id, dataset.info.const_variable, timePeriod)}>
+                        on:click={() => createLayerSimple(dataset.ds_id, dataset.info.const_variable, timePeriod)}>
                       <td></td>
                       <td class="bullet">◦</td>
                       <td colspan="2">{timePeriod}</td>
@@ -529,7 +409,7 @@
                   {/each}
                 {:else}
                   <tr class="layer final" title={dataset.title}
-                      on:click={() => addLayer(
+                      on:click={() => createLayerSimple(
                         dataset.ds_id, dataset.info.const_variable,
                         dataset.info.const_time_period,
                       )}>
