+++ frontend/src/src/layers.js
@@ -3,7 +3,8 @@
 import {get} from 'svelte/store';
 import {layersStore, tasksStore, selectedLayerStore} from './stores.js';
 import {getTask, refreshTask, SUCCESS_STATUS} from './tasks.js';
-import {getDatasetLayerLegend} from './client.js';
+import {getDataset} from './datasets.js';
+import {getDatasetLayerLegend, getDatasetLayerName} from './client.js';
 
 
 // Create a new layer and put it at the top of the stack, so it is displayed last
@@ -49,6 +50,53 @@ export function createLayer(
   layersStore.set(layers);
 
   return layer;
+}
+
+
+// Less customizable version of 'addLayer()'
+export async function createLayerSimple(datasetId, variable, timePeriod) {
+  const dataset = getDataset(datasetId);
+
+  const layerName = await getDatasetLayerName(
+      datasetId, dataset.is_raster, variable, timePeriod,
+  );
+
+  const labels = {
+    primary: null,
+    secondary: null,
+    dataset: null,
+  };
+
+  let title = null;
+
+  if (variable !== null) {
+    labels.primary = variable;
+    labels.dataset = dataset.title;
+
+    if (timePeriod !== null) {
+      labels.secondary = timePeriod;
+      title = variable + '\n\n' + timePeriod + '\n\n' + dataset.title;
+    } else {
+      title = variable + '\n\n' + dataset.title;
+    }
+  } else if (timePeriod !== null) {
+    labels.primary = timePeriod;
+    labels.dataset = dataset.title;
+    title = timePeriod + '\n\n' + dataset.title;
+  } else {
+    labels.primary = dataset.title;
+    title = dataset.title;
+  }
+
+  const layerInfos = {
+    dataset: dataset.ds_id,
+    variable: variable,
+  };
+
+  createLayer(
+      layerName, labels, title, dataset.is_raster, dataset.is_tiled,
+      dataset.info.min_zoom_level, null, layerInfos,
+  );
 }
