+++ frontend/src/src/stores.js
@@ -1,5 +1,8 @@
 import {writable} from 'svelte/store';
 
+export const datasetsStore = writable([]);
+export const datasetTopicsStore = writable([]);
+
 export const areaSelectionStore = writable(null);
 export const areaSelectionLayerStore = writable(null);
